requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13.0
drf-yasg>=1.20.0
django-environ>=0.9.0
psycopg2-binary>=2.9.3
django-filter>=22.1
Faker>=13.3.2